+++ CHANGES.md
@@ -1,5 +1,6 @@
 HEAD
 ---
+* Warn but proxy anyway to processes running outside ringleader's control
 * Add support for environment variable overrides in app config file
 * Support `rbenv` key in config for rbenv-managed projects